requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
pathlib==1.0.1
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
whitenoise==5.3.0
